fix(v19): remediation pass over Codex E1-E20 — real transport, wired life model, contract and ops fixes
Audit findings fixed:
- restore pristine V18.8.1 reference folder (WAV fallback hack reverted; ffmpeg installed instead)
- real aiortc WebRTC ingress + fake_xr_device replaying MP4+pose; honest decode bench
- wire life_model_v19 stage into close-day; weakening/contradicted transitions; 4 handoff-named tests
- v19_visual_context now writes to the real brainlive_v15 tables (shadow schema removed)
- UIIntent.priority int -> float (0..1) across pydantic/schema/C#/delivery_adapter
- outcome watcher: defer calibration labels out of the write transaction (SQLite single-writer)
- GpuArbiter: per-class budgets only for on-demand classes; verify_ollama_unload via /api/ps
- real INSTALL/DOCTOR/RUN/setup_profile scripts; degraded state machine; C# POCOs generated with fields
- G0 timezone fix for test_delivery_feedback (local calendar day vs UTC clock)

Validation: tests/v19 40/40; V18 suite 108/108.

+++ packages/contracts/python/models.py
@@ -34,7 +34,7 @@ TruthLevel = Literal['observed','probable','remembered','inferred','replay']
 class UIIntent(StrictModel):
     ui_intent_id: str; producer: Literal['ultralive','visionrt','brainlive']; source_frame_id: str | None = None
     target_track_id: str | None = None; entity_id: str | None = None; component: str; anchor: dict[str, Any]
-    content: dict[str, Any]; truth_level: TruthLevel; confidence: float; priority: int; ttl_ms: int
+    content: dict[str, Any]; truth_level: TruthLevel; confidence: float; priority: float; ttl_ms: int
     ui_hint: dict[str, Any] = Field(default_factory=dict); evidence_refs: list[str] = Field(default_factory=list); delivery_id: str | None = None
 
 class UIReceipt(StrictModel):
